Hw5 auth

+++ src/services/contacts.js
@@ -2,32 +2,25 @@ import { SORT_ORDER } from "../constants/index.js";
 import { ContactsCollection } from "../db/models/contact.js";
 import { calculatePaginationData } from "../utils/calculatePaginationData.js";
 // console.log("calculatePaginationData => count: ", count, "page: ", page,"perPage: ", perPage);
-// console.log("calculatePaginationData => totalPages: ", totalPages);
 
 export const getAllContacts = async ({
+    userId,
     page = 1,
     perPage = 10,
     sortBy = "_id",
     sortOrder = SORT_ORDER.ASC,
     filter = {}
 }) => {
+    console.log("getAllContacts => userId: ", userId);
     const limit = perPage;
     const skip = (page - 1) * perPage;
-    const contactsQuery = ContactsCollection.find();
+    const contactsQuery = ContactsCollection.find({userId});
     if (filter.type) {
         contactsQuery.where("contactType").equals(filter.type);
     }
     if (filter.isFavourite) {
         contactsQuery.where("isFavourite").equals(filter.isFavourite);
     }
-    // const contactsCount = await ContactsCollection.find()
-    //     .merge(contactsQuery)
-    //     .countDocuments();
-    // const contacts = await contactsQuery
-    //     .skip(skip)
-    //     .limit(limit)
-    //     .sort({ [sortBy]: sortOrder })
-    //     .exec();
 
     const [contactsCount, contacts] = await Promise.all([
         ContactsCollection
@@ -48,10 +41,11 @@ export const getAllContacts = async ({
     };
 };
 
-export const getContactById = async (contactId) => {
-    const contact = await ContactsCollection.findById(contactId);
+export const getContactById = async (userId, contactId ) => {
+    console.log("getContactById => userId: ", userId);
+    const contact = await ContactsCollection.findOne({ userId, _id: contactId });
+    console.log("getContactById => contact: ", contact);
     return contact;
-
 };
 
 export const createContact = async (payload) => {
@@ -59,13 +53,13 @@ export const createContact = async (payload) => {
     return contact;
 };
 
-export const updateContact = async (contactId, payload, options = {}) => {
-    const rawResult = await ContactsCollection.findOneAndUpdate({ _id: contactId }, payload, { new: true, includeResultMetadata: true, ...options });
+export const updateContact = async (userId, contactId, payload, options = {}) => {
+    const rawResult = await ContactsCollection.findOneAndUpdate({ _id: contactId, userId }, payload, { new: true, includeResultMetadata: true, ...options });
     if (!rawResult || !rawResult.value) return null;
     return rawResult.value;
 };
 
-export const deleteContact = async (contactId) => {
-    const contact = await ContactsCollection.findOneAndDelete({ _id: contactId });
+export const deleteContact = async (userId, contactId) => {
+    const contact = await ContactsCollection.findOneAndDelete({ _id: contactId, userId });
     return contact;
 };
